Adds arrow function for 'p' property in 'obj2'
This change introduces an arrow function for the 'p' property in the 'obj2' object. The arrow function correctly references the global object (window object) as it does not have its own 'this'. This modification improves the understanding of the lexical scope behavior for arrow functions in JavaScript.

Fixes #N/A